+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - [+] **@goldfish/react:** add `useMount()` and `useUnmount()`.
 
+### 🐞 Bug Fixes
+
+- [^] **@goldfishjs/react:** keep `this` of the component function in `observer`.
+
 ## 1.1.0
 
 ### 💡 Main Changes